Added renderable object and used namespace 'strata'
Most of the files in /src/core/ are included simply because the
namespace changed, but that's not today's biggest change.

I added a couple of files for terrain management, firstly in /src/core/
(where they will manage the interfacing between the standalone
executable, and its console and scripts, and the terrain-specific code),
and secondly in /src/mesh/. The latter is the place where I will put all
the actual rendering code for the terrain (in the sense of creating
meshes and textures).

I encountered the problem that I needed to add these renderables to the
WorldManager. I considered for a moment to pass addWorldRenderable() and
freeRenderable() as std::function<> objects, but that isn't very elegant
either so for now I simply added the RenderInterface. I think it
shouldn't hurt to include that in a map with code that will heavily
interact with tiny-game-engine code, even though it technically violates
the #include hierarchy that I try to adhere to.

I set up the /src/mesh/ directory such that the main object is a
Terrain, which consists of Layer's. I decided to TypeCluster the layer
such that I can conveniently split layers, add new ones, and delete
obsolete ones, without having to add std::map's to track them and delete
them when necessary. There is no obvious choice for the key so I simply
used a long unsigned int.