Add methods to generate a mesh for a flat layer
It probably should have been easy, but I struggled a bit to get it
right. I wrote a couple of functions and developed a class in order to
generate a triangular mesh consisting of a handful of polygons.

Firstly I created the Vertex and Polygon structs in mesh/element.hpp. I
initially wanted to make them with pointers, but I realised it would
consume too much memory and cause way too much allocation activity, so I
changed my mind and implemented them via an indexed array instead.

The indexed arrays, and the management of lists of Vertex and Polygon
instances, are assigned to the MeshBundle, which by all means and
purposes is a mesh: a collection of vertices plus a set of polygons
linking said vertices into a mesh.

Vertex and polygon objects link back to the index arrays such that they
can be easily removed from there as well. In trying to make everything
efficient, I opted for zero allocations, and coded a hard limit that any
single vertex cannot be part of more than 12 polygons.

The Vertex has a size of 64 bytes (assuming 4 bytes for floats and
integers), and the Polygon is 16 bytes. The difference is that the
Vertex lists all Polygons it belongs to, which is required for
coherence. In this manner, a 1000x1000 mesh with 1m vertices and 2m
polygons would comfortably fit in about 100MB.

The interface of MeshBundle is not complete yet, but it looks like it is
adequate. I have a more or less working generate-flat-mesh function
(createFlatLayer) that generates output that looks reasonable. The next
step is to turn this into a mesh, but I first want to commit the current
state before I have to add even more stuff.

By way of a final remark, I am far from done and I will need tricks of a
considerable degree of complexity to keep memory usage within acceptable
levels while still generating entire mountain chains. The first things
to be done next are turning the MeshBundle into a renderable mesh, and
manipulating that single mesh. Once that works, I'm going to think about
stitching several layers together and try to approach the actual purpose
of this project.