Improve flat layer generation, update per-file license
I wrote functions for generating a flat layer using pretty generic
find-a-vertex functions. It no longer stalls on the ndivs=30 setting,
like before, but it generates 5k polygons where I expected about 2.5k.
So it does seem that there's still a minor issue somewhere.

I also updated the license above each file with an extra line "This file
is part of Chathran Strata: https://github.com/takenu/strata" in order
to connect the source with the project it belongs to. This is suggested
in some of the GPL documentation and I liked the idea.