Finally fixed issues with creating flat layer mesh
One would think nothing is simpler than creating a flat mesh of
equilateral triangles. However, I hope the current exercise is not an
indication of things to come, since I took quite a while to find out
that somehow reinitializing a reference (e.g. "Foo & foo = bar; foo =
bar2") is not at all a guarantee that "&foo == &bar2" will actually be
true. That came as a surprise to me, though in hindsight I should've
known better.

But anyways, for now it is fixed, and I probably will avoid working with
references (as opposed to pointers) from now on, if nothing else simply
to save myself from making silly mistakes that take hours to find. I can
finally make a decent mesh without any double polygons. It's also really
fast, something that makes me very happy since working with hundreds of
thousands of polygons is going to be essential to manipulation of
layered terrains.