Restructure Layer and Stitch, and derive from MeshFragment
I decided that Layers and Stitches are roughly the same kind of object,
with a lot of things in common, so I based them both on a MeshFragment
class. A MeshFragment then is going to be simply a set of polygons on
vertices, with all the rendering-related stuff in it. Meanwhile, the
Layer will specialize in managing large and flat meshes (roughly
2-dimensional) whereas the Stitch will be a long 1-dimensional mesh
which is unusual in that it borrows vertices from adjacent layers.

I still need to work on a suitable way of sharing vertices, but I first
wanted to clean the setup a little, which I hope to have done in this
commit.