Start on classes for stitch meshes
I began with a few classes that should allow implementing stitches.
Stitches are going to be holding everything together, connecting layers
to other layers (horizontally as well as vertically) into a full
multi-layered terrain.

Stitches are not going to contain any vertices (the Layers do that
instead), but they use special constructs to borrow vertices from
Layers. Moreover, Layers are going to un-register vertices they want to
remove through calling purgeVertex on all neighbouring stitches. In this
way we can safely remove all references to a vertex that is no longer
desirable.