Add Mesh base class, remove Junctions
For a unified interface for both Stitches (which work with vertices
belonging to Layers) and Layers (which make up the bulk of the terrain
and use simpler vertex classes than Stitches) I wrote a template Mesh
class to be used for both. The main difference between the types is that
the Stitch variant of vertices is that it needs to also include a
reference to the layer that owns it - and StitchPolygons must also refer
to layers.

I also decided that Junctions are typically unnecessary since extending
Stitches a little bit should be less work than implementing a third type
of Mesh. So they're gone now.

Finally I made a few preparations for using different textures by adding
a resetTexture function to the MeshFragment. I want to make layers one
colour and stitches the other colour, such that one can easily see how
Strata composes its terrains.