Move MeshBundle to new file, use derived Stitch objects
The MeshBundle is too specific to still belong with the Vertex and
Polygon classes, only the Mesh is general enough to 'deserve' being in
the element.hpp file. I am moving things around to specialize the
MeshBundle into a bulk polygon container and the MeshStitch into a
specialized container for, well, line-like meshes that stitch polygons
together.

For consistency I wanted to commit the first round of functional changes
before actually moving element.cpp to bundle.cpp (which will be the next
commit), such that the diff still shows the actual changes (as opposed
to large blocks of code getting moved).

Meanwhile I also changed StitchVertex and StitchPolygon to be
specialised version of Vertex and Polygon. Initially I wanted to make
separate classes but it is actually more straightforward to simply
extend the existing class. I don't want to actually use pointers but it
doesn't feel like a waste to duplicate the position and the array of
polygons, since it helps in writing generic functions in Mesh.

The latter is useful now since Mesh became a template class based on a
VertexType and a PolygonType which are now assumed to at least have the
structure of a Vertex or Polygon. For instance, there needs to be only
one convertToMesh() function now, which would be impractical if there
was no basic structure that could be assumed to exist in the template
parameters.