Move mesh/element.cpp to mesh/bundle.cpp
There are no more classes in element.hpp that define complex functions
that need to be in a separate source file. The basic Vertex and Polygon
classes do not require complex behavior themselves, and the Mesh is a
template class and thus cannot have its larger functions moved into a
separate object file. The remaining mesh/element.cpp is therefore
renamed, following the move of the MeshBundle class to mesh/bundle.hpp
in the previous commit.