Calculate correct normals when creating mesh
Before I used a placeholder tiny::vec3(0.0f,1.0f,0.0f) normal when
determining the normal of a mesh, but I now wrote a slightly more
sophisticated version which calculates the normal based on the cross
product of one of the polygons that the vertex is a part of.

A better version (which I didn't implement) would be to average all
normals required. The reason for not writing that is that a really
correct normal of an edge vertex would need polygons from the layer as
well as from any adjacent stitches, which needs info that will be
available only to more specialised classes (not to the Mesh class
itself). These classes should, in the future, do a re-run over vertices
in each layer in order to remotely set correct normals.