Remove unnecessary ForeignVertex class
When there was still the plan to add a Junction, a ForeignVertex was
used to refer to borrowed vertices (and it did not inherit from Vertex,
either). However, now it's only a complication when we basically need
only a class that is fully a Vertex plus some extra functionality, so
the intermediate is skipped.