Remove Polygon template parameter from Mesh
There seemed to be no more need to distinguish a separate StitchPolygon
since by now a Polygon is really only a triplet of vertex indices plus a
polygon index, which no longer depends on the polygon being in a Stitch
or in a Layer. Consequently, the Mesh no longer needs to use a template
parameter for the Polygon, since everyone will simply use Polygon
itself, not some class derived from it.

On the other hand, the StitchVertex clearly does need to retain its dual
identity with one index (now called remoteIndex) referring to the index
of the vertex in the remote layer, while Vertex::index refers to the
Stitch index.