Move large chunks of MeshBundle to Mesh
I was far too early in moving basically the entire MeshBundle class to a
new file - instead of creating a Mesh that has some of the MeshBundle
members, I should have renamed the MeshBundle as Mesh and moved only the
few really layer-specific functions.

In this commit I mostly move everything 'back' to the Mesh class in
mesh/element.hpp. The move is big enough for Git to consider it a
'rename' rather than a modification of the original file (though I think
this is a bit too coarse). Either way, many things are back to the way
they were.

Meanwhile I also made a small start on functions that find the edge of a
Mesh and another that finds the two most distant points on the edge.
These functions are preparations for the next goal, namely the ability
to split (and eventually, unite) meshes.