Add function to identify edge vertices
Mesh::isEdgeVertex() tests whether a vertex is on the edge by looking
for a vertex that appears only once in its neighbouring polygons.

The algorithm is simple but effective, but it would fail for meshes with
holes or meshes with bottlenecks (hourglass-like features where two
meshes are connected by sharing a single vertex). Since these features
should not normally be useful for Strata, I chose to designate these as
invalid.