Write a findEdgeVertex function for Mesh
The Mesh now has a findEdgeVertex function to start looking for the edge
starting at some random vertex. For a generic mesh there is no way to
deterministically locate an edge vertex (although in principle you could
keep track of one, or all). This function tries to find an edge vertex.

Additionally, I wrote a findRandomEdgeVertex that already makes some
effort to make multiple attempts, but it could be that it gets stuck
(though that should be very, very rare).