Calculate mesh size in analyseShape()
In this commit I include some code for calculating the size of a
MeshFragment, which takes place in the Mesh itself and depends on all
its vertices. I haven't tested it yet, but the code compiles properly.
It isn't used anywhere either, yet, since it takes more before I can
split meshes.