Use std::function to create Layers and Stitches
The Terrain class manages Layers and Stitches, but splitting is a task
best performed by the Layer and Stitch itself, which are far more
closely related to the mesh than the Terrain. Therefore, I decided to
let the Terrain delegate the splitting to the classes themselves, with
the necessary Terrain functionality (i.e. managing MeshFragments)
available through std::function objects.

This is preferable to having to somehow access the entire Terrain object
while splitting meshes, while at the same time avoiding the Terrain
having to do various very specific manipulations on meshes.