Improve vertex interface of Mesh
In this minor commit I make Mesh::addVertex public, and I allow the
owner of the Mesh to convert xVerts into positions. Both functions are
also declared in the Layer to allow redirection. The Layer is starting
to look like a class that ought to be derived from Mesh, but for now the
TypeCluster (containing both stitches and layers) is keeping me from
doing what seems logical.

I will consider splitting the typecluster into stitches and meshes and
deriving the MeshFragment from Mesh itself, either through MeshBundle
and MeshStitch or simply by merging them.

For the splitting, I decided to use an approach where two sublayers are
incrementally grown from the two most-separated vertices. This should
generally result in a decent reduction in maximal end-to-end distance of
any two contained vertices.