Rename MeshBundle -> Bundle and MeshStitch -> Strip
Contrary to the intentions mentioned in the previous commit, I decided
to hold on to the current setup of having both Stitches/Layers and their
Mesh objects, explicitly choosing not to derive these from Mesh. I
renamed the MeshStitch as Strip and the MeshBundle as Bundle as shorter
and more straightforward names. (A Bundle truly bundles vertices and
polygons, whereas a Strip essentially contains only polygons and borrows
its vertices.)

On the other hand, I will aim for a different role: the Layer will
become a true soil layer and can consist of multiple Bundles and Strips,
and the Stitch will connect Layers together at their edges. This also
solves another issue that I became aware of: there needs to be a clear
distinction between stitching together parts of the same layer (just for
the cause of cutting it into manageable parts) and stitching together
the terrain (as in, connecting a layer to the layer below it via a
connecting mesh strip). This distinction will take shape as Strips
versus Stitches, where the latter stitch the terrain from its
constituent layers.

None of this has yet been implemented, in this commit only the names
have been adjusted. In the following commits, I intend focus less on the
reasoning and more on the details of the implementation.