Restructure meshes
This rather massive commit makes deep adjustments to the way meshes are
treated. Now, the Mesh derives from a MeshInterface that is used as a
common class for all meshes, easing unified communication and avoiding
having to distinguish between bundles and strips all over the place. The
MeshInterface in turn derives from the DrawableMesh, which contains the
drawing code.

The Layer was skinned to simply a group of Bundles (for now), with most
functionality moved to Bundle or the Mesh. Splitting will also be done
by the Bundle now, which is far more convenient than having a Layer
manipulating a Mesh from the outside all the time.

The Terrain now manages TypeClusters of Bundles and Strips instead of
Layers and Stitches. It does still contain a vector of Layers but for
now these are not clustered, and their number is expected to be small (a
few dozen or so).

Going forward, splitting will be performed in broadly the same way,
using std::function objects to create new Bundles and Strips (rather
than Layers and Stitches). The Bundle class now owns this function, and
the Strip will later on also implement this function (unless I can
manage to write a generic function, in which case the Mesh gets it).