Manage adjacent meshes in MeshInterface
This commit adds a handful of functions for managing a list of
references to adjacent Mesh objects (a list which I also move to the
MeshInterface in this commit).

Managing adjacent meshes is mandatory for effectively implementing
Stitches, which borrow vertices from other Mesh objects and must be
notified of structural mesh changes (e.g. addition or deletion of
vertices at the edge). Furthermore, adjacent meshes are required for
merging different Mesh objects.