Extend Vertex class by referencing next edge vertex
After a bit of thought I felt the current methods of dealing with mesh
edges are too inefficient, mostly because determining whether a Vertex
is on the edge of a mesh (or not) involves a relatively cumbersome
analysis of the polygons that it is part of. Doing this check once is
fine, but doing it every time we are uncertain about a vertex's
situation (internal or on the edge) seems inefficient.

Additionally, following the edge of a mesh is probably a useful
capability, but it is also quite slow at this time. I therefore added a
variable to the Vertex class which is 0 if the vertex is not on the
edge. For edge vertices, finding the next edge vertex became
straightforward and fast.

Minor adjustments were necessary to allow 0 as the error value, such
that Mesh objects now no longer use vertices[0] and ve[0] as before. I
also moved creation of the error values to the Mesh instead of the
Bundle/Strip classes, since their constructor structure is the same.

Finally I simplified the Vertex construction using C++11 constructor
delegation. Since I already use various C++11-only code elsewhere, I
don't suppose it reduces anyone's ability to compile Strata.