Rename mesh/stitch.hpp -> mesh/strip.hpp
In accordance with the class name change I now also move the file to a
new name. This is particularly important as the Stitch name is already
getting reused in mesh/layer.hpp for a different type of class (intended
for cross-layer stitching, or the special stitches that connect distinct
layers).