Improve edge traversion by using nextEdgeVertex
The functions Mesh::identifyEdgeVertices() and Mesh::checkEdgeVertices()
help in finding the value for Vertex::nextEdgeVertex and verifying its
correctness, respectively. They do this by using a class flag,
mesh::hasDesignatedEdgeVertices. Setting this flag to 'false' after mesh
manipulations is not yet widespread and remains to be done.

After repeated additions the Mesh class has become quite big and hard to
manipulate. I am exploring the possibility of splitting it into two
classes (with one derived from the other) for the sake of reducing the
sheer number of functions in a single class. It is templated, so I
cannot really delegate function bodies to a .cpp file easily, but
nothing seems to prohibit abstracting the more fundamental functionality
into a base class, not because anything else will ever derive from it
but simply because a single class is now doing a real lot of
manipulation --- and I fear there is still a long way to even the most
basic applicability. For all its functions, it still cannot really do
anything besides a few elementary operations on arbitrary meshes. And it
will need to not only analyse meshes, but also modify their shapes, add
and remove vertices and polygons... etcetera.