Split mesh/element.hpp into mesh.hpp and interface.hpp
The MeshInterface is moved to interface.hpp, and the Mesh is moved to
mesh.hpp. Only the Polygon, Vertex, VertPair, the #define for the max
number of polygons a vertex can participate in, and two typedefs remain
in element.hpp. This splitting should help organise the code a bit after
the Mesh is starting to grow beyond any practical size. Next up is
splitting the Mesh itself into as-small-as-possible parts.