Created TopologicalMesh to contain const Mesh funcs
I separated the Mesh into two classes, the Mesh itself (now rather
empty) and the TopologicalMesh, according to whether the functions are
descriptive only (i.e. they leave the mesh's vertices and polygons
unaltered) or mutative (i.e. they alter the mesh itself). This is the
only factoring I could come up with, and it still leaves the
TopologicalMesh as a rather huge class with a few dozen functions and
over three hundred lines of code.

Nonetheless, the code is too intertwined to allow further division, and
the parts that are not intertwined are bound to be so in the future. I
could conceivably create further subdivisions by adding yet another
layer of inheritance, but I think the chain is getting pretty long
already (DrawableMesh -> MeshInstance -> TopologicalMesh -> Mesh ->
Bundle/Strip) and putting the functions in so many classes will just
make development cumbersome in other ways.

For now I have also run out of ideas for restructuring the code, so I
will try to add a bit more and actually get to subdividing flat meshes.