Committing more-or-less-working current version
I used to be having issues with some bug in the splitting functions that
I am working on. I haven't found the bug but the program stopped
crashing for at least one choice of the triangle size. This is the
version I am now committing.

While the bug may not be gone, I will be continuing work during my
holidays when I have only my laptop to work on, hence the commit of not
entirely finished work.