Fixed more stale reference bugs in createFlatLayer()
A rather long while ago (measured not only in time but also in commits)
I had issues with references in what is now Bundle::createFlatLayer().
It turns out I solved the issue only superficially and today I paid the
price for not renaming all of the references (after the point where a
list reallocation potentially invalidates the old references). Indeed,
in fact I am quite stunned that things worked at all as much as they
did.

This commit thus mostly changes a to a2 and b to b2 in bundle.cpp, and
it removes a couple of print-spam statements. Today's commit also shows
that a mesh with half a million vertices can be rendered very smoothly,
so performance issues do seem to be a remote possibility only. (Note
that Strata's aim is to use far smaller meshes of only a few hundred
polygons each, where more precise detail is not explicitly expressed as
terrain relief but instead through parameters applicable to (patches of)
the mesh.

The aim of the near future remains to develop proper mesh-splitting and
mesh-stitching methods, to be followed up by some elementary deformation
functionality.