Complete splitting function
With this commit the last lines are added to
Bundle::splitAddIfNewVertex(). I also added some documentation to the
source on the exact roles of the parameters of each function, to avoid
possible future confusion. Finally I const'ed parameters where
possible/applicable, such that it became harder to mess up by
manipulating the wrong array.

I added an output line that suggests the splitting works fine - although
it would give a bit of a mess for meshes of more typical sizes (I use a
size of only about 30 vertices in this commit).

Next up is the plan to build a Stitch object from the polygons in
between the bundles, to render the new objects instead of the old ones,
and to color the stitches differently.