Add polygons to meshes after splitting
The meshes that result from splitting are assigned polygons by checking
whether all three vertices of the polygon were moved to a single mesh.
If so, a polygon is added.

The main subtlety and core difficulty in this commit is contained in an
adjustment of Mesh::addVertex() where we now also call a newly created
Vertex::clearPolys() function. The latter erases polygon references from
the Vertex. These are initially copied from an initial vertex on copy
construction, but when one actually wants to duplicate parts of a mesh
into a new (but distinct) mesh this results in bad references. On the
other hand, polygon references should be kept if a Mesh is simply copied
as a whole entity (e.g. in case of reallocation) so I don't want to
actually change the behavior of the copy constructor.

The effect is that new, unattached vertices are created when one adds
vertices to a mesh, such that one can be certain that only the position
is kept and the new vertex starts without any polygons that it is a part
of.