Create Strip object when splitting mesh
Today's commit also adds vertices and polygons to the Strip object. I
wrote a single addPolygonWithVertices function in the Strip to allow
adding an entire polygon plus (if necessary) its vertices to it. For
this it needs to know for each vertex what mesh it belongs to, and what
index it has in that mesh. These are the six arguments to the function.

Next in the pipeline is creating graphics meshes for all objects
resulting from the split, and replacing the original mesh with them (as
well as varying the color to see where the stitch lies).