Remove verbatim Selene license statement
Since Selene's source is not included in Strata's repository anymore,
but instead made into a separate library, this commit removes the
license statement that is required for distributions of modified
versions of Selene. The [actual fork of
Selene](https://github.com/takenu/Selene) that is used by Strata does retain the
license.