Create parent layer reference in Mesh
The Layer is going to manage all Bundles and Strips that contain its
polygons, and for that we must link them together. In this commit the
Mesh class (a base to Bundle and Strip) is given a link to this parent
layer. The reverse link (including Bundles and Strips in the Layer) is
partially complete with a vector containing Bundle keys in the layer.
Whether this method is convenient remains to be seen.

In this commit I also dealt with the compiler warning spam that had
accumulated over the last commits.