Remove curious this-argument to class function
I am not sure how it came about or how it wasn't obvious earlier, but
today I realised that the Layer was calling a class function that looked
like a->f(a), i.e. a function of a class with the argument passed to it
the class itself.

Since the function is intended for creating a mesh for the class (the
one owning the function) I don't see any reason why you would want to
call the function on one instance in order to create a mesh for another,
so I did some cleaning up.