Replace old mesh by new mesh fragments when splitting
In this commit the split has become visible due to discarding the old
Bundle (and its associated mesh) and creating meshes for the new Bundles
(which were already created in earlier commits but not yet visible).

The Strip fragment is not yet visible, I'm leaving creating a mesh for
it as an exercise for a future commit.