Avoid adding vertices with 0 polygons to a Bundle
The initial setup of the splitting function was a bit too simple, and it
did not check that every vertex added to a Bundle was actually going to
be properly connected to that Bundle (resulting in a valid mesh). The
result was vertices without adjacent polygons that were, so to speak,
purely living on the edge.

Such meshes are considered invalid (since it invalidates certain basic
assumptions on mesh topologies) and must be avoided. Therefore, this
commit makes some adjustments to guarantee that vertices are only added
to Bundles if the resulting Bundle is a valid mesh.

However, it is not complete because it does not solve another issue that
I foresee: that of vertices that happen to be poorly located at the end
of a split, which are only part of 1 polygon for which the two
neighbouring vertices are part of two distinct Bundles.

I want to solve that issue by merging the outlier vertex with one of the
Bundles (swapping is complex and wouldn't work anyways for very spiky
meshes with lots of such outlier vertices), but that is for another
commit. This problem doesn't manifest itself right now (yet, but I am
sure it's there) so I will need to do a bit more splitting to verify
that my intended solution actually works.