Create strip.cpp for Strip functions, move neighbor check
In this commit I moved the complex 'are-the-vertices-of-the-farthest-
pair-really-far-enough-apart' check to TopologicalMesh, since I realised
the Strip splitting would need it as well (and also it is a purely
topological question that one might want to ask more frequently in the
future).

I also created a new file for the more complex functions of the Strip
class, which will be used in the future for the splitting functions.