Move splitting code to Mesh
Much of the splitting that I was doing could be generalised rather
easily and put into the Mesh class (which is a base to both the Strip
and the Bundle). This should aid splitting Strips as well as Bundles
without duplicating code extensively.

The original Bundle::split() function still remains but it is relatively
short, and all related splitting functions were moved to the Mesh.
Additionally, some functionality from the Strip was moved to the Mesh.

All of this is making the Mesh a rather large-to-behold class with some
lengthy function bodies in it. On the bright side, I don't foresee a lot
more methods on Mesh in the foreseeable future, and in the end there are
only a handful of manipulations - split, join, refine, coarsen - that
one could intend to do on a Mesh.