Split Strip into two shorter Strips
This commit finishes Strip::split() using the methods previously moved
from the Bundle into the Mesh. Only a minor adjustment of existing
methods was necessary for this (namely adding a Vertex in a compatible
way in Mesh::splitMesh()).

The splitting works now, but only once, because it turns out that there
is no code yet to adjust a Strip's references when its adjacent Bundle
is split (or ceases to exist for some other reason). Addressing this
issue will be the topic of further commits.