Split mesh multiple times and cleanup vert/poly arrays
Matters were a bit more subtle than anticipated in the previous commit
message, because the adjacentMeshes-structure was not yet used and
deleting polygons or vertices from a Mesh was not yet possible.

This commit corrects these omissions and adds functions for merging
vertices, as well as for deleting vertices and polygons and correcting
references in adjacent meshes (i.e. Strips).

Deleting and merging is rather dangerous and I already put several
operations private in Mesh to avoid accidental abuse. However, more
safeguards are likely to be useful and sanity checks to verify a Mesh's
consistency would be welcome, too.