Fix bugs in Mesh
When scaling up the mesh a bit (to 8 divisions) a few bugs appeared,
which are fixed in this commit. The first was about the error vertex at
(0,0,0) being considered a valid vertex by addIfNewVertex(), resulting
in polygons using this invalid vertex.

The second and third were cleaning up the polygon/vertex arrays in the
wrong order in deleteVertexFromArray and deletePolygonFromArray. I
actually wouldn't have expected the previous commit to appear to work
smoothly with a bug like this, but anyways, it is gone now.

After this commit I am going to scale up a bit more, as well as making
some additional precautions that merging edge vertices works even if
there are long 'tails' attached to the mesh. (The current merge solution
works for a single spike, but it may work poorly if there is a long tail
attached to the mesh. Furthermore it does not guarantee that the max
link limit is honoured, such that merging two polygons with (combined)
too many links would break the mesh.