Begin work on proper tracking of adjacent meshes
I scaled up a bit more and divided some more, and while nothing horrible
happens there does turn out to be a lot of shortcomings with the
adjacent mesh management.

This commit does nearly nothing and contains mostly garbage, but since I
did spend a few hours trying to figure out a convenient way to make
adjacency more complete I took a moment to write down my current
thoughts on the matter.

Adjacent meshes are tracked because mesh modifications (of a Bundle,
since a Strip doesn't have vertices of its own) must be communicated
such that references can be updated. Every strip knows the owning Bundle
and the remote index of the vertex in that Bundle for each of its
borrowed vertices. This structure works only if the Bundle will let the
Strip know whenever the Vertex ceases to exist, and if the Strip knows
which Bundle is the one that owns its vertex (preferably without having
to lookup its reference in the Bundle type cluster).

Currently a lot of the adjacent mesh references are not refreshed when
splitting. The references to the old mesh fragment are automatically
removed by the MeshInterface destructor, but re-adding them (especially
re-adding only when necessary) is more tricky.

This commit contains a commented-away unfinished attempt to fix this by
writing a single function in the MeshInterface. Coded but not yet
functional, there is a more promising attempt via a Strip-specific
function in the MeshInterface (to be used upon Bundle splitting) one
way, and an as of yet undeveloped Strip member function that, when
splitting, carefully looks which of its adjacent meshes need to be
passed on to which of its successors.

Concluding, the following needs to be done:
- Strip::updateRemoteVertexIndices() should be put to use such that Strip
vertices are adjusted correctly, and Bundle::split() needs to call
MeshInterface::updateVerticesInAdjacentMeshes
- Strip needs to be given a function for selecting which adjacent meshes
  are required for which of its successor Strip objects, and then use
MeshInterface::addAdjacentMesh() (with the second argument 'true') to
set the correct adjacency dependencies.