Continued on functions for establishing adjacency
In this commit I played around a bit with the way in which I am trying
to figure out whether a mesh is adjacent to another mesh. It is not
quite complete yet, but at least I got everything to compile and run
normally so that I can continue later.

Next up is reviewing if the new functions look reliable enough and
testing that the correct adjacencies are found. After that, resolving
the spiky/long-tail splitting issue should be the main priority.