Start alternative adjacency tracking
The function updateVerticesInAdjacentMeshes() that I was writing was
basically only useful for replacing Bundles by other Bundles, and
nonsensical otherwise. The increasing amount of clutter necessary to
avoid the confusion (such as empty function bodies for
Bundle::updateRemoteVertexIndices()) is telling me that maybe I am going
all wrong about it and the adjacency should be far more explicit:
Bundles are adjacent to Strips which borrow their vertices, and Strips
are adjacent to Bundles since they need the vertices of the Bundles.

Trying to make the adjacency abstract simply doesn't work because the
Strips only need to tell the Bundles they exist (and the Bundles will
never know how the Strips work internally, and they do not have any
access to their structure) whereas the Bundles need to explain to the
Strips how their edge vertices are evolving, whether edges are being
split, and whether vertices are being merged.

This inequality must be handled by allowing very specfic operations on
one object originating at the other, and making a dozen pure virtual
functions in the interface simply for 'communication' is a pointless and
messy exercise that only makes the code totally chaotic and likely
inefficient as well.

I therefore started developing new adjacency in the Strip and the
Bundle, and I intend to more or less move all related functions from the
MeshInterface to the Bundle and the Strip, leaving the MeshInterface
purely an interface and not a means of communication (and perhaps the
name ought to imply this limited functionality anyways, *nomen est omen*).

The next commits therefore are going to be about redoing the adjacency,
hopefully this time leading to a more elegant and effective way of
tracking adjacency through various episodes of splitting meshes.