Remove obsolete adjacency code
Given that everything was known to work well even with totally
inadequate adjacency, I decided to first remove the old methods from the
MeshInterface. The MeshInterface is now again a very thin class (perhaps
too thin but we shall see), and the Bundle and Strip (mostly the latter)
are going to do the work.

I already wrote a bit more functionality in the Strip to handle
adjacency, more is expected in upcoming commits.