Add function for checking adjacency completeness
To make sure that adjacency is established properly when splitting
meshes it is useful to have a function that checks whether the links and
references have been established in a sensible manner. The function
Strip::isAdjacencyComplete() does this check and can in principle be
used after every mesh modification to verify that all vertices of the
strip have correct vertices.

The code is as yet untested. Also, I would like to develop a function
that checks polygon coherence - that is, whether polygon edges on the
outside of the Strip are also edges of some polygon of the Bundle. This
equates to checking that there exists a polygon that contains both
vertices.