Add edge splitting and edge swapping
I was tempted to summarize this commit as "add an enormous heap of crap"
which would have been quite straight-to-the-point. But there are
actually a few good things here, so let's get to these.

Firstly, I needed over a dozen hours over two weeks to 'improve' strata
into something that can do half what the previous committed version
could. That is because I replaced the older merge-problem-vertices
approach by a split-problem-polygons approach. Roughly, this means that
instead of changing vertex positions, we add a new vertex on an edge,
which is cleaner.

In particular, merging excessively would very quickly
lead to problems with the max number of allowable polygons per vertex.
In the new situation, it is far less likely to hit this limit since two
vertices (one of which is the spike vertex) get one extra polygon each.
To avoid even problems related to this, I also wrote functions for
rebalancing the mesh's links, with the purpose of swapping polygon edges
when this helps polygons be more equilateral or when it reduces the
number of polygons on a vertex with many such links.

That was the good news. For the bad news, I think the logic determining
edge swaps is flawed, but after I fixed other issues there is no more
edge swapping so right now nothing seems wrong. Therefore the edge
swapping should be reviewed later, when it is actually getting used.

More bad news is that some of the meshes are not anymore getting
rendered, and since I am running out of time for this weekend I commit
it anyways since at least everything compiles and runs.

But the main message of this commit is that we can now split edges by
adding a vertex midway (while also replacing the polygons next to the
edge that is split), and that we have some kind of a start for how to
swap edges. Both are supported by a lot of additional functionality for
analysing and modifying the topology in fundamental ways.

Finally, the issues that occurred confirm that there is a need for
strict tests on topological correctness to be developed. The future
focus is therefore to find the remaining issues in the splitting (now
that we use edge splitting rather than vertex merging), to track
adjacency better (that should be relatively minor), and to start using
existing tests (and add more tests) to enforce that every step involving
topological modifications leaves us with a valid topology.